Add FormState wrapper for CustomDropdown and rename DropdownItem to CustomDropdownItem
This commit will include a widget that wraps a CustomDropdown in a FormState, by doing this we gain the ability to call verify and onSave  methods when we use the dropdown inside a Form.
Rename the dropdown item class from DropdownItem to CustomDropdownItem, so the name is inline with the project and doesn't make any confusion.